+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/ui/films/FilmsListViewModel.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.viewmodels
+package com.sw.sw_api_kotlin_project.ui.films
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
