+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/ui/planets/PlanetsFragment.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.fragments
+package com.sw.sw_api_kotlin_project.ui.planets
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,46 +6,48 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.sw.sw_api_kotlin_project.adapters.FilmsAdapter
+import com.sw.sw_api_kotlin_project.adapters.PlanetsAdapter
 import com.sw.sw_api_kotlin_project.base.BaseFragment
-import com.sw.sw_api_kotlin_project.databinding.FragmentFilmsBinding
+import com.sw.sw_api_kotlin_project.databinding.FragmentPlanetsBinding
 import com.sw.sw_api_kotlin_project.repository.APIRepository
-import com.sw.sw_api_kotlin_project.viewmodels.FilmsListViewModel
-import com.sw.sw_api_kotlin_project.viewmodels.FilmsListViewModelFactory
 
-class FilmsFragment : BaseFragment() {
-    private lateinit var viewModel: FilmsListViewModel
-    private var _binding: FragmentFilmsBinding? = null
+
+class PlanetsFragment : BaseFragment() {
+
+    private lateinit var viewModel: PlanetsViewModel
+    private var _binding: FragmentPlanetsBinding? = null
     private val binding get() = checkNotNull(_binding)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         viewModel = ViewModelProvider(
             this,
-            FilmsListViewModelFactory(APIRepository())
-        )[FilmsListViewModel::class.java]
+            PlanetsViewModelFactory(
+                APIRepository()
+            )
+        )[PlanetsViewModel::class.java]
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentFilmsBinding.inflate(inflater, container, false)
+        _binding = FragmentPlanetsBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        viewModel.fetchFilms()
+        viewModel.fetchPlanets()
         observeApiLoadingEvent(viewModel)
     }
 
     override fun addObservers() {
         super.addObservers()
-        viewModel.films.observe(viewLifecycleOwner) {
-            val adapter = FilmsAdapter(it!!.results)
-            binding.filmRecycler.adapter = adapter
-            binding.filmRecycler.layoutManager = LinearLayoutManager(context)
+        viewModel.planet.observe(viewLifecycleOwner) {
+            val adapter = PlanetsAdapter(it!!.results)
+            binding.planetsRecycler.adapter = adapter
+            binding.planetsRecycler.layoutManager = LinearLayoutManager(context)
         }
     }
 
@@ -53,5 +55,6 @@ class FilmsFragment : BaseFragment() {
         _binding = null
         super.onDestroy()
     }
+
 
 }
