+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/ui/planets/PlanetsViewModel.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.viewmodels
+package com.sw.sw_api_kotlin_project.ui.planets
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
