+++ SW_API_Kotlin_Project/app/src/main/res/layout/fragment_planets.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="fragment_planets"
-            type="com.sw.sw_api_kotlin_project.viewmodels.PlanetsViewModel" />
+            type="com.sw.sw_api_kotlin_project.ui.planets.PlanetsViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
