+++ SW_API_Kotlin_Project/app/src/main/res/layout/item_films.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="fragment_people"
-            type="com.sw.sw_api_kotlin_project.viewmodels.FilmsListViewModel" />
+            type="com.sw.sw_api_kotlin_project.ui.films.FilmsListViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:tools="http://schemas.android.com/tools"
