+++ SW_API_Kotlin_Project/app/src/main/res/navigation/navigation_graph.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_people"
-        android:name="com.sw.sw_api_kotlin_project.fragments.PeopleFragment"
+        android:name="com.sw.sw_api_kotlin_project.ui.people.PeopleFragment"
         android:label="PeopleFragment"
         tools:layout="@layout/fragment_people_list" />
 
     <fragment
         android:id="@+id/nav_films"
-        android:name="com.sw.sw_api_kotlin_project.fragments.FilmsFragment"
+        android:name="com.sw.sw_api_kotlin_project.ui.films.FilmsFragment"
         android:label="FilmsFragment"
         tools:layout="@layout/fragment_films" />
 
     <fragment
         android:id="@+id/nav_planets"
-        android:name="com.sw.sw_api_kotlin_project.fragments.PlanetsFragment"
+        android:name="com.sw.sw_api_kotlin_project.ui.planets.PlanetsFragment"
         android:label="PlanetsFragment"
         tools:layout="@layout/fragment_planets" />
